update avn with extension_builder changes
