+++ pallets/avn/src/tests/session_handler_tests.rs
@@ -1,6 +1,6 @@
 // Copyright 2022 Aventus Network Services (UK) Ltd.
 
-use crate::mock::{extension_builder::ExtBuilder, *};
+use crate::mock::*;
 use sp_avn_common::event_types::Validator;
 use sp_runtime::testing::UintAuthorityId;
